+++ packages/auth/__tests__/auth-unit-test.ts
@@ -1643,7 +1643,7 @@ describe('auth unit test', () => {
                             { Name: 'phone_number', Value: 'phone_number' },
                             { Name: 'email_verified', Value: 'false' },
                             { Name: 'phone_number_verified', Value: 'true' },
-                            { Name: 'sub', Value: 'fefefe' }
+                            { Name: 'sub', Value: '123-456789' }
                         ]);
                     });
                 });
@@ -1666,7 +1666,8 @@ describe('auth unit test', () => {
                     email: 'email',
                     phone_number: 'phone_number',
                     email_verified: false,
-                    phone_number_verified: true
+                    phone_number_verified: true,
+                    sub: "123-456789"
                 }
             });
 
@@ -1965,15 +1966,19 @@ describe('auth unit test', () => {
                     return callback('err', null);
                 });
 
-            expect.assertions(1);
+            const spyon3 = jest.spyOn(CognitoUser.prototype, 'getUserData');
+
+            expect.assertions(2);
             try {
                 await auth.currentUserPoolUser();
             } catch (e) {
                 expect(e).toBe('err');
+                expect(spyon3).not.toBeCalled();
             }
 
             spyon.mockClear();
             spyon2.mockClear();
+            spyon3.mockClear();
         });
 
         test('get user data error because of user is deleted or disabled', async () => {
@@ -2041,7 +2046,6 @@ describe('auth unit test', () => {
             spyon2.mockClear();
             spyon3.mockClear();
         });
-    });
     });
 
     describe('sendCustomChallengeAnswer', () => {