+++ packages/aws-amplify-angular/src/aws-amplify-angular.module.ts
@@ -1,6 +1,20 @@
+/*
+ * Copyright 2017-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
+ * the License. A copy of the License is located at
+ *
+ *     http://aws.amazon.com/apache2.0/
+ *
+ * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
+ * and limitations under the License.
+ */
+
 import { NgModule , forwardRef} from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+// tslint:disable:max-line-length
 import { AuthenticatorComponent } from './components/authenticator/authenticator/authenticator.factory';
 import { AuthenticatorComponentCore } from './components/authenticator/authenticator/authenticator.component.core';
 import { ConfirmSignInComponent } from './components/authenticator/confirm-sign-in-component/confirm-sign-in.factory';
@@ -27,7 +41,7 @@ import { ChatBotComponent } from './components/interactions/chatbot/chatbot.fact
 import { ChatbotComponentCore } from './components/interactions/chatbot/chatbot.component.core'; 
 import { DynamicComponentDirective } from './directives/dynamic.component.directive';
 import { FormComponent } from './components/common/form.component';
-
+// tslint:enable:max-line-length
 
 const components = [
   AuthenticatorComponent,
@@ -55,8 +69,7 @@ const components = [
   ChatBotComponent,
   ChatbotComponentCore,
   FormComponent,
-]
-
+];
 
 @NgModule({
   imports: [