+++ packages/aws-amplify-angular/src/components/storage/s3-album-component/s3-album.component.ionic.ts
@@ -2,17 +2,18 @@ import { Component, Input, ViewEncapsulation, Injector, ElementRef } from '@angu
 import { AmplifyService, AuthState } from '../../../providers';
 import { S3AlbumComponentCore } from './s3-album.component.core';
 
-const template = `
-
-<div class="amplify-album-container">
-  <amplify-s3-image-ionic
-    class="amplify-image-container"
-    *ngFor="let item of list"
-    path="{{item.path}}"
-    (selected)="onImageSelected($event)"
-  ></amplify-s3-image-ionic>
+const template =  `
+<div class="amplify-album">
+  <div class="amplify-album-container">
+    <amplify-s3-image-core
+      class="amplify-image-container"
+      *ngFor="let item of list"
+      path="{{item.path}}"
+      (selected)="onImageSelected($event)"
+    ></amplify-s3-image-core>
+  </div>
 </div>
-`
+`;