+++ packages/aws-amplify-react-native/CHANGELOG.md
@@ -3,6 +3,234 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.1-unstable.2"></a>
+## [2.0.1-unstable.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.0...aws-amplify-react-native@2.0.1-unstable.2) (2018-09-05)
+
+
+### Bug Fixes
+
+* **aws-amplify-react-native:** fix the link in the requireNewPassword component ([bdd816a](https://github.com/aws/aws-amplify/commit/bdd816a))
+
+
+
+
+<a name="2.0.1-unstable.1"></a>
+## [2.0.1-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.0...aws-amplify-react-native@2.0.1-unstable.1) (2018-08-30)
+
+
+### Bug Fixes
+
+* **aws-amplify-react-native:** fix the link in the requireNewPassword component ([bdd816a](https://github.com/aws/aws-amplify/commit/bdd816a))
+
+
+
+
+<a name="2.0.0"></a>
+# [2.0.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.26...aws-amplify-react-native@2.0.0) (2018-08-28)
+
+
+### Features
+
+* UI Components ([1ff1abd](https://github.com/aws/aws-amplify/commit/1ff1abd))
+
+
+### BREAKING CHANGES
+
+* UI Components
+
+
+
+
+<a name="1.0.7-unstable.26"></a>
+## [1.0.7-unstable.26](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.25...aws-amplify-react-native@1.0.7-unstable.26) (2018-08-28)
+
+
+* Amplify ui migration (#1517) ([41d3184](https://github.com/aws/aws-amplify/commit/41d3184)), closes [#1517](https://github.com/aws/aws-amplify/issues/1517)
+
+
+### BREAKING CHANGES
+
+* UI Components
+
+
+
+
+<a name="1.0.7-unstable.25"></a>
+## [1.0.7-unstable.25](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.24...aws-amplify-react-native@1.0.7-unstable.25) (2018-08-27)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.24"></a>
+## [1.0.7-unstable.24](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.23...aws-amplify-react-native@1.0.7-unstable.24) (2018-08-27)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.23"></a>
+## [1.0.7-unstable.23](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.21...aws-amplify-react-native@1.0.7-unstable.23) (2018-08-27)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.22"></a>
+## [1.0.7-unstable.22](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.21...aws-amplify-react-native@1.0.7-unstable.22) (2018-08-25)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.21"></a>
+## [1.0.7-unstable.21](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.20...aws-amplify-react-native@1.0.7-unstable.21) (2018-08-24)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.20"></a>
+## [1.0.7-unstable.20](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.19...aws-amplify-react-native@1.0.7-unstable.20) (2018-08-24)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.19"></a>
+## [1.0.7-unstable.19](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.17...aws-amplify-react-native@1.0.7-unstable.19) (2018-08-24)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.18"></a>
+## [1.0.7-unstable.18](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.17...aws-amplify-react-native@1.0.7-unstable.18) (2018-08-24)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.17"></a>
+## [1.0.7-unstable.17](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.15...aws-amplify-react-native@1.0.7-unstable.17) (2018-08-24)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.16"></a>
+## [1.0.7-unstable.16](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.15...aws-amplify-react-native@1.0.7-unstable.16) (2018-08-24)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.15"></a>
+## [1.0.7-unstable.15](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.14...aws-amplify-react-native@1.0.7-unstable.15) (2018-08-24)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.14"></a>
+## [1.0.7-unstable.14](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.13...aws-amplify-react-native@1.0.7-unstable.14) (2018-08-24)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.13"></a>
+## [1.0.7-unstable.13](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.11...aws-amplify-react-native@1.0.7-unstable.13) (2018-08-23)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.12"></a>
+## [1.0.7-unstable.12](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.11...aws-amplify-react-native@1.0.7-unstable.12) (2018-08-23)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.11"></a>
+## [1.0.7-unstable.11](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.10...aws-amplify-react-native@1.0.7-unstable.11) (2018-08-23)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.10"></a>
+## [1.0.7-unstable.10](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.9...aws-amplify-react-native@1.0.7-unstable.10) (2018-08-23)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.9"></a>
+## [1.0.7-unstable.9](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.8...aws-amplify-react-native@1.0.7-unstable.9) (2018-08-23)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.8"></a>
+## [1.0.7-unstable.8](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.7...aws-amplify-react-native@1.0.7-unstable.8) (2018-08-22)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.7"></a>
+## [1.0.7-unstable.7](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.6...aws-amplify-react-native@1.0.7-unstable.7) (2018-08-22)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.6"></a>
+## [1.0.7-unstable.6](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.5...aws-amplify-react-native@1.0.7-unstable.6) (2018-08-21)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.5"></a>
+## [1.0.7-unstable.5](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.4...aws-amplify-react-native@1.0.7-unstable.5) (2018-08-21)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+<a name="1.0.7-unstable.4"></a>
+## [1.0.7-unstable.4](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.3...aws-amplify-react-native@1.0.7-unstable.4) (2018-08-20)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react-native
+
 <a name="1.0.7-unstable.3"></a>
 ## [1.0.7-unstable.3](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@1.0.7-unstable.2...aws-amplify-react-native@1.0.7-unstable.3) (2018-08-19)